+++ db/migrate/20211028204634_create_workout_routes.rb
@@ -5,9 +5,9 @@ class CreateWorkoutRoutes < ActiveRecord::Migration[6.1]
       t.belongs_to :workout
       t.belongs_to :route
       t.int :repetitions
-      t.float :distance
-      t.float :pace
-      t.float :duration
+      t.decimal :distance
+      t.decimal :pace    # minutes/mile as a decimal. So 8.5 = 8 minutes 30 seconds per mile.
+      t.decimal :duration
       t.int :heart_rate
       t.text :description
